backend/requirements.txt
altgraph==0.17
apyori==1.1.2
astroid==2.3.3
auto-py-to-exe==2.6.6
bottle==0.12.18
bottle-websocket==0.2.9
braintree==4.0.0
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9
defusedxml==0.6.0
Django==2.0.2
django-allauth==0.41.0
django-cors-headers==3.2.1
django-countries==6.1.2
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-payments==0.13.0
django-payu==0.5
django-rest-auth==0.9.5
djangorestframework==3.11.0
Eel==0.11.0
future==0.18.2
gevent==1.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
isort==4.3.21
joblib==0.14.1
kmodes==0.10.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
pefile==2019.4.18
Pillow==7.0.0
pycparser==2.20
pycryptodome==3.9.7
PyInstaller==3.6
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
pywin32-ctypes==0.2.0
pywix==0.2.1
requests==2.22.0
requests-oauthlib==1.3.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
stripe==2.44.0
suds-jurko==0.6
typed-ast==1.4.1
urllib3==1.25.8
whichcraft==0.6.1
wincertstore==0.2
wrapt==1.11.2
xmltodict==0.12.0
